Add logistic regression prediction methods
This commit adds methods related to making predictions for a data set, using a pre-fitted logistic regression model. The user would instantiate a logistic regression object by giving it a vector of coefficients and an intercept. The logistic regression object then can be used for making predictions, calculating confidence and probability.